src/cloud-functions/parameter_publisher/requirements.txt
google-cloud-pubsub==2.15.2